Move forms from mozorg to base app, incl tests
